Fix an error caused by a diff having a whitespace line after the @@...@@ divider. This error was most likely caused by the splitlines() command in the diff creation seeing a newline at the start of the string and acting in an unforeseen way.
The fix strips all newline characters from the lines, which means that we no longer have to use .text[1:], which may cause errors down the line. After the newlines are stripped, a check is performed to see if it is the first line in the new group. if it is, the text is instead converted to a tab character, preserving this whitespace line. 

Additionally, changed error reporting in views and template mismatch.

git-svn-id: https://vis.cs.umd.edu/svn/projects/predict/prod@60653 943b2781-5fa1-c61d-a95c-b09a2b48979a